Step3: add basic auto-configuration
Add the most basic auto-configuration we could find. It creates a
ConnectionFactory on the local broker if none exists and if HornetQ is
available.

We also make sure to run this auto-config *before* the general JMS
auto-configuration so that the JmsTemplate can be created properly.
We also add a test to make sure that everything is working as expected.

You can either run this demo with the regular 1.2.3.RELEASE by adding an
exclude of the existing HornetQ support in Boot (see DemoApplication) or
you can checkout the snicoll/auto-config-demo branch and build that
locally to have a 1.2.4.BUILD-SNAPSHOT that does not include temporarily
the official Spring Boot support.